requirements.txt
accelerate==0.22.0
altgraph==0.17.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dl-translate==0.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.3
fsspec==2023.9.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
Pillow==9.3.0
prettytable==3.8.0
protobuf==4.24.2
psutil==5.9.5
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
sacremoses==0.0.53
safetensors==0.3.3
selenium==4.12.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tinydb==4.8.0
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
transformers==4.33.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
wsproto==1.2.0
zipp==3.16.2
